Added camera

+++ src/peacefulotter/game/Maths/Quaternion.java
@@ -2,9 +2,9 @@ package peacefulotter.game.Maths;
 
 public class Quaternion
 {
-    private double x, y, z, w;
+    private float x, y, z, w;
 
-    public Quaternion(double x, double y, double z, double w )
+    public Quaternion(float x, float y, float z, float w )
     {
         this.x = x;
         this.y = y;
@@ -12,26 +12,27 @@ public class Quaternion
         this.w = w;
     }
 
-    public double getX() { return x; }
-    public void setX( double x ) { this.x = x; }
+    public float getX() { return x; }
+    public void setX( float x ) { this.x = x; }
 
-    public double getY() { return y; }
-    public void setY( double y ) { this.y = y; }
+    public float getY() { return y; }
+    public void setY( float y ) { this.y = y; }
 
-    public double getZ() { return z; }
-    public void setZ( double z ) { this.z = z; }
+    public float getZ() { return z; }
+    public void setZ( float z ) { this.z = z; }
 
-    public double getW() { return w; }
-    public void setW( double w ) { this.w = w; }
+    public float getW() { return w; }
+    public void setW( float w ) { this.w = w; }
 
-    public double length()
+    public float length()
     {
-        return Math.sqrt( BMaths.square( x ) + BMaths.square( y ) + BMaths.square( z ) + BMaths.square( w ) );
+        return (float) Math.sqrt(
+                BMaths.square( x ) + BMaths.square( y ) + BMaths.square( z ) + BMaths.square( w ) );
     }
 
     public Quaternion normalize()
     {
-        double length = length();
+        float length = length();
         return new Quaternion( x / length, y / length, z / length, w  / length );
     }
 
@@ -42,19 +43,19 @@ public class Quaternion
 
     public Quaternion mul( Quaternion other )
     {
-        double x_ = x * other.getW() + w * other.getX() + y * other.getZ() - z * other.getY();
-        double y_ = y * other.getW() + w * other.getY() + z * other.getX() - x * other.getZ();
-        double z_ = z * other.getW() + w * other.getZ() + x * other.getY() - y * other.getX();
-        double w_ = w * other.getW() - x * other.getX() - y * other.getY() - z * other.getZ();
+        float x_ = x * other.getW() + w * other.getX() + y * other.getZ() - z * other.getY();
+        float y_ = y * other.getW() + w * other.getY() + z * other.getX() - x * other.getZ();
+        float z_ = z * other.getW() + w * other.getZ() + x * other.getY() - y * other.getX();
+        float w_ = w * other.getW() - x * other.getX() - y * other.getY() - z * other.getZ();
         return new Quaternion( x_, y_, z_, w_ );
     }
 
     public Quaternion mul( Vector3f other )
     {
-        double x_ =  w * other.getX() + y * other.getZ() - z * other.getY();
-        double y_ =  w * other.getY() + z * other.getX() - x * other.getZ();
-        double z_ =  w * other.getZ() + x * other.getY() - y * other.getX();
-        double w_ = -x * other.getX() - y * other.getY() - z * other.getZ();
+        float x_ =  w * other.getX() + y * other.getZ() - z * other.getY();
+        float y_ =  w * other.getY() + z * other.getX() - x * other.getZ();
+        float z_ =  w * other.getZ() + x * other.getY() - y * other.getX();
+        float w_ = -x * other.getX() - y * other.getY() - z * other.getZ();
         return new Quaternion( x_, y_, z_, w_ );
     }
 }
